+++ default.nix
@@ -3,7 +3,7 @@
 }:
 mkDerivation {
   pname = "agda-unused";
-  version = "0.1.1";
+  version = "0.2.0";
   src = ./.;
   isLibrary = true;
   isExecutable = true;